+++ pay-api/src/pay_api/models/payment_line_item.py
@@ -29,12 +29,12 @@ class PaymentLineItem(db.Model, BaseModel):
     filing_fees = db.Column(db.Integer, nullable=False)
     fee_schedule_id = db.Column(db.Integer, ForeignKey('fee_schedule.fee_schedule_id'), nullable=False)
     quantity = db.Column(db.Integer, nullable=True)
-    processing_fees = db.Column(db.Integer, nullable=True)
-    service_fees = db.Column(db.Integer, nullable=True)
+    processing_fees = db.Column(db.Float, nullable=True)
+    service_fees = db.Column(db.Float, nullable=True)
     description = db.Column(db.String(200), nullable=True)
-    gst = db.Column(db.Integer, nullable=True)
-    pst = db.Column(db.Integer, nullable=True)
-    total = db.Column(db.Integer, nullable=False)
+    gst = db.Column(db.Float, nullable=True)
+    pst = db.Column(db.Float, nullable=True)
+    total = db.Column(db.Float, nullable=False)
 
     @classmethod
     def find_by_id(cls, identifier: int):
